ok i thihnk all stuff is implemented now but i am  scared to run it TT

+++ PaymentStrategy.java
@@ -1,4 +1,4 @@
 public interface PaymentStrategy {
 
-    public void pay(int n );
+    public void pay(double n );
 }
